server/requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
openai==0.28.0
python-dotenv==1.0.1
gunicorn==20.1.0